+++ CHANGELOG.md
@@ -6,8 +6,10 @@
 
 ### New features
 
+* Add `kala.control.OptionContainer::getOrThrow()`
 * Add `kala.collection.base.OrderedTraversable::mapIndexed(CollectionFactory<U, ?, R>, IndexedFunction<T, U>)`
 * Add `kala.collection.base.OrderedTraversable::mapIndexedTo(G, IndexedFunction<T, U>)`
+* Add `kala.collection.MapLike::getOrThrow(K)`
 * Add `kala.collection.Set::added(E)`
 * Add `kala.collection.Set::addedAll(E...)`
 * Add `kala.collection.Set::addedAll(Iterable<E>)`